Moved unit tests to main-library folder
