--- old/package.json
@@ -1,15 +1,0 @@
-{
-  "scripts": {
-    "dev": "tsc --watch src/index.ts",
-    "build": "tsc src/index.ts && uglifyjs src/index.js --output src/index.js",
-    "test": "jest"
-  },
-  "devDependencies": {
-    "@types/jest": "^29.1.0",
-    "@types/node": "^18.7.23",
-    "jest": "^29.1.1",
-    "ts-jest": "^29.0.3",
-    "typescript": "^4.8.4",
-    "uglify-js": "^3.17.2"
-  }
-}
