--- old/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "target": "es5",
-    "module": "commonjs",
-    "esModuleInterop": true,
-    "moduleResolution": "node",
-    "forceConsistentCasingInFileNames": true
-  }
-}
