+++ packages/main-library/src/__tests__/index.test.ts
@@ -1,6 +1,5 @@
 import { read as readBufferWorkBook } from "xlsx"
-import jsonxlsx from "../index"
-import { IContent, IJsonSheet, ISettings } from "../../types"
+import jsonxlsx, { IContent, IJsonSheet, ISettings } from "../index"
 
 describe("json-as-xlsx", () => {
   it("should return undefined if sheets array is empty", () => {
